requirements.txt
# Core framework
Django>=4.2

# Machine Learning / Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Scientific stack
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Image processing & visualization
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Deployment (optional for later)
gunicorn>=21.2.0
whitenoise>=6.5.0
